requirements.txt
frozendict==2.3.2
matplotlib==3.1.2
more_itertools==4.2.0
numpy==1.17.4
pyedifice==0.0.10
pyqtgraph==0.11.0rc0
PySide2==5.15.2.1
